# expression: implement vectorized evaluation for builtinCastTimeAsTimeSig