PCP #12101

<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

This PR implements vectorized `makebuiltinCastTimeAsTimeSig`.

### What is changed and how it works?

```sh
➜ make vectorized-bench VB_FILE=Cast VB_FUNC=builtinCastTimeAsTimeSig
cd ./expression && \
	go test -v -benchmem \
		-bench=BenchmarkVectorizedBuiltinCastFunc \
		-run=BenchmarkVectorizedBuiltinCastFunc \
		-args "builtinCastTimeAsTimeSig"
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinCastFunc/builtinCastTimeAsTimeSig-VecBuiltinFunc-12         	   26919	     44802 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastTimeAsTimeSig-NonVecBuiltinFunc-12      	   17143	     70504 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastTimeAsTimeSig-VecBuiltinFunc#01-12      	   17361	     69502 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastTimeAsTimeSig-NonVecBuiltinFunc#01-12   	   12207	     99565 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	7.719s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test