@ekalinin 2019-11-08
/run-all-tests